requirements_gpu.txt
# 基础依赖
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# 数据库
psycopg2-binary>=2.9.0

# 机器学习和GPU支持
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
scikit-learn>=1.3.0

# 搜索和网络
duckduckgo-search>=4.0.0
requests>=2.31.0

# OpenAI
openai>=1.0.0

# 工作流
pocketflow>=0.1.0

# 日志和工具
tqdm>=4.65.0
colorama>=0.4.6 